♻️ [refactor] #152 : types 활용
